refactor: move tests into CleanAction `mod.rs` file
